pyproject.toml
[project]
name = "repl-nix-budgetwise"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.40.1",
]
